feature: Add message receiver for News's Keyword
This commit adds a message receiver for News's Keyword" using RabbitMQ. The process involves first receiving the message (containing the Keywords' event data) from the message broker queue, registering it to the Keywords table, and enabling the News aggregate to access this data.

The data obtained from Keywords is used during the registration of News and also the retrieval of News data along with its keywords.